Collision between sprites
> Made player use the DoPhysics() method from Sprite.cs for better maintainability and consistency

> Both objects now collide and both are knocback-ed when they touch

> Collision with the top now doesn't result in a crazy velocity this was due to how the Gravity attribute was used to add velocity downwards. See amendments in Sprite.cs line 516.

+++ Sprites/Mob.cs
@@ -27,9 +27,9 @@ namespace Bound.Sprites
         {
             var distance = (_game.Player.Position - Position).Length();
 
-            if (_game.Player.Position.X < Position.X)
+            if ((int)_game.Player.Position.X < (int)Position.X)
                 Velocity -= new Vector2(_speed, 0);
-            else if (_game.Player.Position.X > Position.X)
+            else if ((int)_game.Player.Position.X > (int)Position.X)
                 Velocity += new Vector2(_speed, 0);
         }
     }
